% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Book Recommender Server Logic}
\usage{
app_server(input, output, session, books, ratingmat, model)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{books}{Book metadata}

\item{ratingmat}{User-book rating matrix}

\item{model}{Trained recommender model}
}
\description{
Book Recommender Server Logic
}
