% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_book_data.R
\name{load_book_data}
\alias{load_book_data}
\title{Load and Clean Book Data}
\usage{
load_book_data()
}
\value{
A list containing cleaned books, ratings, and ratingmat.
}
\description{
This function loads and preprocesses book and rating data.
}
