% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train Recommender Model}
\usage{
train_model(ratingmat)
}
\arguments{
\item{ratingmat}{A realRatingMatrix of user ratings.}
}
\value{
A recommender model object.
}
\description{
Builds a user-based collaborative filtering model.
}
